% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{barplotn}
\alias{barplotn}
\title{Drawing a figure like barplot()}
\usage{
barplotn(formula, data = NULL, ..., las = 1, cex.axis = 1.1,
  cex.lab = 1.3, font.lab = 2, lwd.bor = 2, lwd.axis = 1,
  lwd.0 = 1, lwd.stat = 1, lty.0 = 3, col.fill = c("#1F77B47F",
  "#FF7F0E7F", "#2CA02C7F", "#D627287F", "#9467BD7F", "#8C564B7F",
  "#E377C27F", "#7F7F7F7F", "#BCBD227F", "#17BECF7F"),
  col.bor = c("#1F77B4FF", "#FF7F0EFF", "#2CA02CFF", "#D62728FF",
  "#9467BDFF", "#8C564BFF", "#E377C2FF", "#7F7F7FFF", "#BCBD22FF",
  "#17BECFFF"), col.stat = "default", col.0 = "#0000007F",
  length = "auto", space = 0.5, names = NULL, xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, SE = F, SD = F,
  horizontal = F, beside = F, legend = F, pos.leg = "outright",
  pch.leg = 22, bty.leg = "n", bg.leg = "transparent",
  pt.cex.leg = 2, tx.cex.leg = 1.1, pt.col.leg = NULL,
  pt.bg.leg = NULL, tx.col.leg = NULL, leg.lab = NULL,
  leg.sp = 2.5, inset = 1, leg.title = NULL, tit.col.leg = NULL,
  mar = c(2, 3.8, 1, 1), mgp = c(2.5, 0.5, 0), tcl = -0.2,
  inversion = F, inv.col = "#FFFFFF")
}
\arguments{
\item{formula}{Data, e.g. numeric vector, formula, e.g. y ~ x, or other object containing analysis result}

\item{data}{If formula is inputted in "formula" parameter, a data.frame (or list) from which the variables in formula should be taken.}

\item{...}{Argument to be passed to methods. Please see barplot().}

\item{las}{las, defauls is 1}

\item{cex.axis}{axis cex, default is 1.1}

\item{cex.lab}{label cex, default is 1.3}

\item{font.lab}{label font size, default is 2}

\item{lwd.bor}{box border lwd, default is 2}

\item{lwd.axis}{axis lwd, default is 1}

\item{lwd.0}{Zero line lwd, default is 1}

\item{lwd.stat}{Error bar lwd, default is 1}

\item{lty.0}{Line type of zero line, default is 3}

\item{col.fill}{fill color}

\item{col.bor}{border color}

\item{col.stat}{Mean and error bar color}

\item{col.0}{Zero line color}

\item{length}{Length of vertical bar of tip in error bar}

\item{space}{Barplot space. Default is 0.5}

\item{names}{names}

\item{xlim}{x limit}

\item{ylim}{y limit}

\item{xlab}{x label}

\item{ylab}{y label}

\item{SE}{If set "T", standard error is drawn. Default is "F".}

\item{SD}{If set "T", standard deviation is drawn. Default is "F".}

\item{horizontal}{horizontal, default is "F".}

\item{beside}{beside}

\item{legend}{If legend is needed, set "T". Default is "F".}

\item{pos.leg}{Legend position. In addition to position of legend(), "outtopright, "outright", "outbottomright" and "outbottom" are able to select. Default is "outright".}

\item{pch.leg}{Legend pch, default is 22.}

\item{bty.leg}{Legend box type. Default is ""n.}

\item{bg.leg}{Legend background}

\item{pt.cex.leg}{Points cex in legend, default is 2.}

\item{tx.cex.leg}{Text cex in legend, default is 1.1.}

\item{pt.col.leg}{Points color in legend.}

\item{pt.bg.leg}{Points background color in legend.}

\item{tx.col.leg}{Text color in legend.}

\item{leg.lab}{Legend label}

\item{leg.sp}{Legend space, default is 2.5.}

\item{inset}{Legend inset, default is 1.}

\item{leg.title}{Legend title}

\item{tit.col.leg}{Legend title color}

\item{mar}{mar, default is c(2,3.8,1,1).}

\item{mgp}{mgp, default is c(2.5,0.5,0).}

\item{tcl}{tcl, default is -0.2.}

\item{inversion}{Inversion mode. If set "T", plot is drawn with inversion color. Default is "F".}

\item{inv.col}{Inversion color, if set inversion = "T". Default is "#FFFFFF".}
}
\description{
Drawing a figure like barplot()
}
