% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{boxplotn}
\alias{boxplotn}
\title{Drawing a figure like boxplot()}
\usage{
boxplotn(x = NULL, formula = NULL, data = NULL, ..., las = 1,
  cex.axis = 1.1, cex.lab = 1.3, font.lab = 2, lty = 1,
  outline = F, all = T, jitter.method = "jitter", staplelwd = NA,
  boxwex = 0.5, xlab = NULL, ylab = NULL, names = NULL,
  xlim = NULL, ylim = NULL, notch = F, horizontal = F,
  xaxt = "s", yaxt = "s", col.fill = NULL, col.bor = NULL,
  col.dot = NULL, col.stat = NULL, col.bg = "#FFFFFF",
  pch.dot = 16, pch.stat = 21, cex.dot = 0.7, cex.stat = 1,
  lwd.bor = 1, lwd.stat = 1, lwd.dot = 1, noise = 1,
  side = "center", side.sp = 0.3, reflect = T, Mean = F, SE = F,
  SD = F, legend = F, pos.leg = "outright", pch.leg = 22,
  bty.leg = "n", bg.leg = "transparent", pt.cex.leg = 2,
  tx.cex.leg = 1.1, pt.col.leg = NULL, pt.bg.leg = NULL,
  tx.col.leg = NULL, leg.lab = NULL, leg.sp = 2.5, inset = 1,
  leg.title = NULL, tit.col.leg = NULL, mar = c(2, 3.8, 1, 1),
  mgp = c(2.5, 0.5, 0), tcl = -0.2, add = F, inversion = F,
  inv.col = "#FFFFFF")
}
\arguments{
\item{x}{Data, e.g. numeric vector, formula, e.g. y ~ x, or other object containing analysis result}

\item{formula}{formula}

\item{data}{If formula is inputted in "x" or "formula" parameter, a data.frame (or list) from which the variables in formula should be taken.}

\item{...}{Argument to be passed to methods. Please see boxplot().}

\item{las}{las, defauls is 1}

\item{cex.axis}{axis cex, default is 1.1}

\item{cex.lab}{label cex, default is 1.3}

\item{font.lab}{label font size, default is 2}

\item{lty}{lty}

\item{outline}{If set "T", outliners are drawn. Default is "F".}

\item{all}{If set "T", all points are drawn. Default is "T".}

\item{jitter.method}{how to draw jitter, "jitter", "swarm", "center", "hex" and "square" are able to select. Default is "jitter".}

\item{staplelwd}{staplelwd, default is "NA".}

\item{boxwex}{boxwex, default is 0.5.}

\item{xlab}{x label}

\item{ylab}{y label}

\item{names}{names}

\item{xlim}{x limit}

\item{ylim}{y limit}

\item{notch}{notch, default is "F".}

\item{horizontal}{horizontal, default is "F".}

\item{xaxt}{xaxt, default is "s".}

\item{yaxt}{yaxt, default is "s".}

\item{col.fill}{fill color}

\item{col.bor}{border color}

\item{col.dot}{points color}

\item{col.stat}{Mean and error bar color}

\item{col.bg}{background color}

\item{pch.dot}{points pch, default is 16}

\item{pch.stat}{mean points pch, default is 21}

\item{cex.dot}{points cex, default is 0.7}

\item{cex.stat}{mean points cex, default is 1}

\item{lwd.bor}{box border lwd, default is 1}

\item{lwd.stat}{mean and error bar lwd, default is 1}

\item{lwd.dot}{points lwd, default is 1}

\item{noise}{scatter level of points, default is 1}

\item{side}{move direction of boxplot, "left", "center" and "right" are able to select. Default is "center".}

\item{side.sp}{Magnitude of move of boxplot, default is 0.3.}

\item{reflect}{If set "T", points are not drawn in next to boxplot and reflected. Default is "T".}

\item{Mean}{If set "T", mean points are drawn. Default is "F".}

\item{SE}{If set "T", standard error is drawn. Default is "F".}

\item{SD}{If set "T", standard deviation is drawn. Default is "F".}

\item{legend}{If legend is needed, set "T". Default is "F".}

\item{pos.leg}{Legend position. In addition to position of legend(), "outtopright, "outright", "outbottomright" and "outbottom" are able to select. Default is "outright".}

\item{pch.leg}{Legend pch, default is 22.}

\item{bty.leg}{Legend box type. Default is ""n.}

\item{bg.leg}{Legend background}

\item{pt.cex.leg}{Points cex in legend, default is 2.}

\item{tx.cex.leg}{Text cex in legend, default is 1.1.}

\item{pt.col.leg}{Points color in legend.}

\item{pt.bg.leg}{Points background color in legend.}

\item{tx.col.leg}{Text color in legend.}

\item{leg.lab}{Legend label}

\item{leg.sp}{Legend space, default is 2.5.}

\item{inset}{Legend inset, default is 1.}

\item{leg.title}{Legend title}

\item{tit.col.leg}{Legend title color}

\item{mar}{mar, default is c(2,3.8,1,1).}

\item{mgp}{mgp, default is c(2.5,0.5,0).}

\item{tcl}{tcl, default is -0.2.}

\item{add}{If set "T", boxplot is able to overdrawn on previous boxplot. Default is "F".}

\item{inversion}{Inversion mode. If set "T", plot is drawn with inversion color. Default is "F".}

\item{inv.col}{Inversion color, if set inversion = "T". Default is "#FFFFFF".}
}
\description{
Drawing a figure like boxplot()
}
