% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{category.axis}
\alias{category.axis}
\title{Drawing categorized axis}
\usage{
category.axis(main, sub, data = NULL, main.axis.at = NULL,
  main.axis.length = 3, sub.axis.at = NULL, lwd = 1, bar.lwd = 1,
  cex.axis = 1.1, las = 1, x.intsp = 0.6, y.intsp = 1.8,
  horizontal = F, mar = c(3.8, 3.8, 1, 1), mgp = c(2.5, 0.5, 0),
  tcl = -0.2, inversion = F, inv.col = "#FFFFFF")
}
\arguments{
\item{main}{Main category, this is given as vector (e.g. c("S", "R"))}

\item{sub}{Sub category, this is given as vector (e.g. c("1", "10", "100"))}

\item{data}{If formula is inputted in "formula" parameter, a data.frame (or list) from which the variables in formula should be taken.}

\item{main.axis.at}{Drawing position of main axis}

\item{main.axis.length}{Bar length of main axis}

\item{sub.axis.at}{Drawing position of sub axis}

\item{lwd}{sub axis lwd, default is 1.}

\item{bar.lwd}{main axis lwd, default is 1.}

\item{cex.axis}{axis cex, default is 1.1.}

\item{las}{las}

\item{x.intsp}{Inter space of main axis bar, defauit is 0.6.}

\item{y.intsp}{Inter space of main and sub axis, default is 1.8,}

\item{horizontal}{horizontal, default is "F".}

\item{mar}{mar, default is c(3.8,3.8,1,1).}

\item{mgp}{mgp, default is c(2.5,0.5,0).}

\item{tcl}{tcl, default is -0.2.}

\item{inversion}{Inversion mode. If set "T", plot is drawn with inversion color. Default is "F".}

\item{inv.col}{Inversion color, if set inversion = "T". Default is "#FFFFFF".}
}
\description{
Drawing categorized axis
}
