% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{category_axis}
\alias{category_axis}
\title{Drawing categorized axis used with overdraw()}
\usage{
category_axis(main, sub, data = NULL, main.axis.at = NULL,
  main.axis.length = 3, sub.axis.at = NULL, lwd = 1, bar.lwd = 1,
  cex.axis = 1.1, las.main = 1, las.sub = 1, x.intsp = 0.6,
  y.intsp = 1.8, horizontal = F)
}
\arguments{
\item{main}{Main category, this is given as vector (e.g. c("S", "R")) or column name of data}

\item{sub}{Sub category, this is given as vector (e.g. c("1", "10", "100")) or column name of data}

\item{data}{a data.frame}

\item{main.axis.at}{Drawing position of main axis}

\item{main.axis.length}{Bar length of main axis}

\item{sub.axis.at}{Drawing position of sub axis}

\item{lwd}{sub axis lwd, default is 1.}

\item{bar.lwd}{main axis lwd, default is 1.}

\item{cex.axis}{axis cex, default is 1.1.}

\item{las.main}{las of main category}

\item{las.sub}{las of sub category}

\item{x.intsp}{Inter space of main axis bar, defauit is 0.6.}

\item{y.intsp}{Inter space of main and sub axis, default is 1.8,}

\item{horizontal}{horizontal, default is "F".}
}
\description{
Drawing categorized axis used with overdraw()
}
\examples{
d <- data.frame(x = c(1:10, 11:20, 21:30, 31:40),
                group = rep(c("A","B","A", "B"), each = 10),
                treatment = rep(c("X","Y"), each = 20))
boxplotn(x ~ group + treatment, data = d, xaxt = "n",
         xlab = "", mar = c(3.8, 3.8, 1, 1))
overdraw(category_axis(main = "treatment", sub = "group",
                       data = d))

boxplotn(x ~ group + treatment, data = d, xaxt = "n",
         xlab = "", mar = c(3.8, 3.8, 1, 1))
overdraw(category_axis(main = c("X", "Y"), sub = c("A", "B")))

}
\seealso{
[plotn::overdraw]
}
