% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{cmd_make}
\alias{cmd_make}
\title{Make command consist of plotn and overdraw to store a figure as a object}
\usage{
cmd_make(...)
}
\arguments{
\item{...}{expression() object or character of plot command}
}
\description{
Make command consist of plotn and overdraw to store a figure as a object
}
\examples{
g1 <- cmd_make(expression(plotn(1:50)),
               expression(overdraw(abline(v = 30),
                                   abline(h = 20),
                                   points(1:10 + 1, 10:1))))

d <- data.frame(x = c(1:10, 11:20, 21:30, 31:40),
                group = rep(c("A","B","A", "B"), each = 10),
                treatment = rep(c("X","Y"), each = 20))
g2 <- cmd_make(expression(boxplotn(x ~ group + treatment, data = d, xaxt = "n",
                                   xlab = "", mar = c(3.8, 3.8, 1, 1))),
               expression(overdraw(category.axis(main = "treatment", sub = "group",
                       data = d))))

}
\seealso{
[plotn::cmd_plot]
}
