% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{cmd_plot}
\alias{cmd_plot}
\title{Make figure using command made by cmd_make}
\usage{
cmd_plot(command)
}
\arguments{
\item{command}{plot command made by cmd_make()}
}
\description{
Make figure using command made by cmd_make
}
\examples{
g1 <- cmd_make(expression(plotn(1:50)),
               expression(overdraw(abline(v = 30),
                                   abline(h = 20),
                                   points(1:10 + 1, 10:1))))
cmd_plot(g1)

}
\seealso{
[plotn::cmd_make]
}
