% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{col_genelator}
\alias{col_genelator}
\title{Genelating dafault color in plotn}
\usage{
col_genelator(palette = "d3", palette_types = NULL, number = 10,
  alpha = 1)
}
\arguments{
\item{palette}{Color palette, "default", "ggplot2" and palettes in ggsci}

\item{palette_types}{"Continuous" or "discrete" in ggplot2 palette or palette types in ggsci}

\item{number}{Number of colors to generate}

\item{alpha}{Transparency}
}
\description{
Genelating dafault color in plotn
}
\examples{
cols <- col_genelator(palette = "npg")
cols <- col_genelator(palette = "uchicago", palette_type = "light")

}
\seealso{
[ggsci] <https://cran.r-project.org/web/packages/ggsci/vignettes/ggsci.html>
}
