% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{histn}
\alias{histn}
\title{Drawing a figure like hist()}
\usage{
histn(x = NULL, formula = NULL, data = NULL, ..., xlab = NULL,
  ylab = NULL, las = 1, main = "", cex.axis = 1.1, cex.lab = 1.3,
  font.lab = 2, col.fill = NULL, col.bor = NULL, hist.dens = NA,
  hist.ang = 45, kernel = F, freq = T, col.line = NULL,
  col.ker = "transparent", ker.dens = NA, ker.ang = 45,
  lwd.hist = 1, lwd.line = 2, breaks = NULL, horizontal = F,
  legend = F, pos.leg = "outright", pch.leg = 22, bty.leg = "n",
  bg.leg = "transparent", pt.cex.leg = 2, tx.cex.leg = 1.1,
  pt.col.leg = NULL, pt.bg.leg = NULL, tx.col.leg = NULL,
  leg.lab = NULL, leg.sp = 2.5, inset = 1, leg.title = NULL,
  tit.col.leg = NULL, mar = c(3.8, 3.8, 1, 1), mgp = c(2.5, 0.5, 0),
  tcl = -0.2, inversion = F, inv.col = "#FFFFFF")
}
\arguments{
\item{x}{Data, e.g. numeric vector, formula, e.g. y ~ x, or other object containing analysis result}

\item{formula}{formula}

\item{data}{If formula is inputted in "x" or "formula" parameter, a data.frame (or list) from which the variables in formula should be taken.}

\item{...}{Argument to be passed to methods. Please see hist().}

\item{xlab}{x label}

\item{ylab}{y label}

\item{las}{las, defauls is 1}

\item{main}{Main title}

\item{cex.axis}{axis cex, default is 1.1}

\item{cex.lab}{label cex, default is 1.3}

\item{font.lab}{label font size, default is 2}

\item{col.fill}{histogram fill color}

\item{col.bor}{histogram border color}

\item{hist.dens}{Density of histgram fill}

\item{hist.ang}{Angle of histgram fill stripe, default is 45 degree.}

\item{kernel}{If set "T", density curve is also drawn. Default is "F".}

\item{freq}{If set "T", data is transformed into frequency data. Default is "F".}

\item{col.line}{density curve line color}

\item{col.ker}{density curve fill color}

\item{ker.dens}{Density of density curve fill}

\item{ker.ang}{Angle of density curve fill stripe, default is 45 degree.}

\item{lwd.hist}{histgram lwd, default is 1}

\item{lwd.line}{density curve lwd, default is 2}

\item{breaks}{breaks}

\item{horizontal}{horizontal, default is "F".}

\item{legend}{If legend is needed, set "T". Default is "F".}

\item{pos.leg}{Legend position. In addition to position of legend(), "outtopright, "outright", "outbottomright" and "outbottom" are able to select. Default is "outright".}

\item{pch.leg}{Legend pch, default is 22.}

\item{bty.leg}{Legend box type. Default is ""n.}

\item{bg.leg}{Legend background}

\item{pt.cex.leg}{Points cex in legend, default is 2.}

\item{tx.cex.leg}{Text cex in legend, default is 1.1.}

\item{pt.col.leg}{Points color in legend.}

\item{pt.bg.leg}{Points background color in legend.}

\item{tx.col.leg}{Text color in legend.}

\item{leg.lab}{Legend label}

\item{leg.sp}{Legend space, default is 2.5.}

\item{inset}{Legend inset, default is 1.}

\item{leg.title}{Legend title}

\item{tit.col.leg}{Legend title color}

\item{mar}{mar, default is c(3.8,3.8,1,1).}

\item{mgp}{mgp, default is c(2.5,0.5,0).}

\item{tcl}{tcl, default is -0.2.}

\item{inversion}{Inversion mode. If set "T", plot is drawn with inversion color. Default is "F".}

\item{inv.col}{Inversion color, if set inversion = "T". Default is "#FFFFFF".}
}
\description{
Drawing a figure like hist()
}
