% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{is.formula}
\alias{is.formula}
\title{Check which object is formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{objects}
}
\description{
Check which object is formula
}
\examples{
x <- z ~ y
is.formula(x)

}
