% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{leap.year}
\alias{leap.year}
\title{Judging leap year}
\usage{
leap.year(year)
}
\arguments{
\item{year}{The start year which experiments were conducted (e.g. 1999, 2001...).}
}
\description{
Judging leap year
}
