% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{Mean.pt}
\alias{Mean.pt}
\title{Drawing mean points}
\usage{
Mean.pt(formula, data = NULL, at = NULL, SD = F, SE = F,
  group = NULL, mean.column = 1, pch = 21, cex = 1.5,
  col.mean = "#000000", col.bg = "#FFFFFF", lwd.mean = 1,
  col.bar = "#000000", lwd.bar = 1, lty = 1, length = 0.5,
  horizontal = F, plot = T, mar = c(2, 3.8, 1, 1))
}
\arguments{
\item{formula}{Data, e.g. numeric vector, formula, e.g. y ~ x, or other object containing analysis result}

\item{data}{If formula is inputted in "formula" parameter, a data.frame (or list) from which the variables in formula should be taken.}

\item{at}{Drawing position}

\item{SD}{If set "T", standard deviation is drawn. Default is "F".}

\item{SE}{If set "T", standard error is drawn. Default is "F".}

\item{group}{Grouping factor}

\item{mean.column}{Column which means are stored.}

\item{pch}{pch, default is 21.}

\item{cex}{cex, default is 1.5.}

\item{col.mean}{mean points color, default is "#000000".}

\item{col.bg}{mean points background color, default is "#FFFFFF".}

\item{lwd.mean}{mean lwd, default is 1.}

\item{col.bar}{error bar color, default is "#000000".}

\item{lwd.bar}{error bar lwd, default is 1.}

\item{lty}{line type of error bar, default is 1.}

\item{length}{Length of vertical bar of tip in error bar, default is 0.5}

\item{horizontal}{horizontal, default is "F".}

\item{plot}{If set "F", calculate is only done.}

\item{mar}{mar, default is c(2,3.8,1,1).}
}
\description{
Drawing mean points
}
