% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{month.axis}
\alias{month.axis}
\title{Drawing a month labeled axis}
\usage{
month.axis(leap = F, period = 1, year = NULL, start = c(1, 1),
  lwd = 1, month.lab = "a", cex.axis = 1.1, las = 1, mar = c(3.8,
  3.8, 1, 1), mgp = c(2.5, 0.5, 0), tcl = -0.2, inversion = F,
  inv.col = "#FFFFFF")
}
\arguments{
\item{leap}{If set "T", a year is treated as leap year.}

\item{period}{Periods (years) which experiments were conducted.}

\item{year}{The start year which experiments were conducted (e.g. 1999, 2001...).}

\item{start}{The start month and day, e.g. October 15 is c(10,15).}

\item{lwd}{lwd}

\item{month.lab}{Month label, "a", "n", "i" and "f" are able to select. "a" is abbreviation, "n" is number, "i" is initial and "f" is full.}

\item{cex.axis}{axis cex, default is 1.1,}

\item{las}{las}

\item{mar}{mar, default is c(2,3.8,1,1).}

\item{mgp}{mgp, default is c(2.5,0.5,0).}

\item{tcl}{tcl, default is -0.2.}

\item{inversion}{Inversion mode. If set "T", plot is drawn with inversion color. Default is "F".}

\item{inv.col}{Inversion color, if set inversion = "T". Default is "#FFFFFF".}
}
\description{
Drawing a month labeled axis
}
