% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{month_axis}
\alias{month_axis}
\title{Drawing a month labeled axis used with overdraw()}
\usage{
month_axis(
  leap = F,
  period = 1,
  year = NULL,
  start = c(1, 1),
  lwd = 1,
  month.lab = "a",
  cex.axis = 1.1,
  las = 1
)
}
\arguments{
\item{leap}{If set "T", a year is treated as leap year.}

\item{period}{Periods (years) which experiments were conducted.}

\item{year}{The start year which experiments were conducted (e.g. 1999, 2001...).}

\item{start}{The start month and day, e.g. October 15 is c(10,15).}

\item{lwd}{lwd}

\item{month.lab}{Month label, "a", "n", "i" and "f" are able to select. "a" is abbreviation, "n" is number, "i" is initial and "f" is full.}

\item{cex.axis}{axis cex, default is 1.1,}

\item{las}{las}
}
\description{
Drawing a month labeled axis used with overdraw()
}
\examples{
d1 <- data.frame(Date = 1:100, x = rnorm(100, 1, 1))
#This data starting at January 1st, 2004
plotn(d1, line = TRUE, pch = NA, xaxt = "n", xlab = "Month")
overdraw(month_axis(period = 1, year = 2004, start = c(1, 1)))

d2 <- data.frame(Date = 1:365, x = rnorm(365, 1, 1))
#This is treated as 365 days data starting at March 25th, 2019
plotn(d2, line = TRUE, pch = NA, xaxt = "n", xlab = "Month")
overdraw(month_axis(period = 2, year = 2019, start = c(3, 25),
                    month.lab = "i"))

}
\seealso{
[plotn::overdraw]
}
