% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{overdraw}
\alias{overdraw}
\title{Function of overdrawing of low level plot function on plot function in plotn library}
\usage{
overdraw(...)
}
\arguments{
\item{...}{low level plot function, e.g. points(...) etc. Multiple functions are able to be set.}
}
\description{
Function of overdrawing of low level plot function on plot function in plotn library
}
\examples{
plotn(1:50)
overdraw(abline(v = 30), abline(h = 20), points(1:10 + 1, 10:1))

}
