% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{par.set}
\alias{par.set}
\title{Function of overdrawing of low level plot function on plot function in plotn library}
\usage{
par.set(x, mar = c(3.8, 3.8, 1, 1), mgp = c(2.5, 0.5, 0), tcl = -0.2,
  inversion = F, inv.col = "#FFFFFF")
}
\arguments{
\item{x}{low level plot function, e.g. points()...}

\item{mar}{mar, default is c(3.8,3.8,1,1).}

\item{mgp}{mgp, default is c(2.5,0.5,0).}

\item{tcl}{tcl, default is -0.2.}

\item{inversion}{Inversion mode. If set "T", plot is drawn with inversion color. Default is "F".}

\item{inv.col}{Inversion color, if set inversion = "T". Default is "#FFFFFF".}
}
\description{
Function of overdrawing of low level plot function on plot function in plotn library
}
