% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{plotn}
\alias{plotn}
\title{Drawing a figure like plot()}
\usage{
plotn(x = NULL, y = NULL, mode = "s", ...)
}
\arguments{
\item{x}{Data, e.g. numeric vector, formula, e.g. y ~ x, or other object containing analysis result}

\item{y}{If numeric vector is inputted in "formula" parameter, numeric vector is also inputted in y}

\item{mode}{Plotting mode. Setting "s" is single group plot, while setting "m" is multiple groups plot. Default is "s".}

\item{...}{Argument to be passed to methods. Please see plot().}
}
\description{
Drawing a figure like plot()
}
