% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{print.plotn}
\alias{print.plotn}
\title{Make figure using command made by plotn_object()}
\usage{
## Default S3 method:
}
\arguments{
\item{x}{plotn object made by plotn_object()}

\item{...}{further arguments passed to or from other methods.}

\item{plot}{If set "F", only print command as a character vector}
}
\description{
Make figure using command made by plotn_object()
}
\examples{
n1 <- plotn_object(plotn(1:50),
        overdraw(abline(v = 30),
                 abline(h = 20),
                 points(1:10 + 1, 10:1))
      )
n1

}
\seealso{
[base::print][plotn::plotn_object]
}
