% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotn.R
\name{theme_change}
\alias{theme_change}
\title{Color theme change}
\usage{
theme_change(
  default_col = NULL,
  default_fill = NULL,
  palette = "d3",
  palette_types = NULL,
  number = 10,
  col.alpha = 1,
  fill.alpha = 0.5
)
}
\arguments{
\item{default_col}{Default setting colors which is used for dot, line and border colors}

\item{default_fill}{Default setting colors which is used for fill colors}

\item{palette}{Color palette, "default", "ggplot2" and palettes in ggsci. Please see col_genelator().}

\item{palette_types}{"Continuous" or "discrete" in ggplot2 palette or palette types in ggsci}

\item{number}{Number of colors to generate}

\item{col.alpha}{Transparency of default color}

\item{fill.alpha}{Transparency of default fill color}
}
\description{
Color theme change
}
\examples{
d <- data.frame(x = c(1:10, 11:20, 21:30), group = rep(c("A","B","C"), each = 10))
theme_change()
boxplotn(x ~ group, data = d)

theme_change(palette = "uchicago", palette_type = "light")
boxplotn(x ~ group, data = d)

}
\seealso{
[ggsci] <https://cran.r-project.org/web/packages/ggsci/vignettes/ggsci.html>

[plotn::col_genelator]
}
